feat: client can request tensors after data arrived
When setting enable_request_task_data_tensors=True, the client will send
a federated event to the server requesting the tensors to be sent.

It means that the server can first send the task data without tensors
and wait for the client to requet the tensors to be sent.

It should allow supporing the ClientAPI in the future.